refactor: revise mobile responsiveness code and modify project folder sturcture
