+++ frontend/src/app/settings/profile/page.tsx
@@ -49,7 +49,7 @@ const ProfileEditPage: React.FC = () => {
         </button>
     );
 
-    const Items = [
+    const items = [
         'Account',
         'Authentication',
         'Organizations',
@@ -103,7 +103,7 @@ const ProfileEditPage: React.FC = () => {
                         <span style={{ fontWeight: '500' }}>Profile</span>
                     </div>
                     {
-                        Items.map((item, index) => {
+                        items.map((item, index) => {
                             return <div key={index}>
                                 <div className="p-2 w-full cursor-pointer hover:bg-slate-400 hover:text-black">
                                     <span >{item}</span>
@@ -135,7 +135,7 @@ const ProfileEditPage: React.FC = () => {
                         {fileList.length >= 1 ? null : uploadButton}
                     </Upload>
                     <Modal open={previewOpen} title={previewTitle} footer={null} onCancel={handleCancel}>
-                        <img alt="example" style={{ width: '100%' }} src={previewImage} />
+                        <Image alt="example" style={{ width: '100%' }} src={previewImage} />
                     </Modal>
 
                     <button className='gap-2 border-[rgb(229 231 235)] w-[20%] inline-flex cursor-pointer items-center justify-center rounded-[0.5rem] border-[1px] p-1 hover:scale-105' onClick={handleSaveUserProfile}>
