+++ frontend/src/app/workSpace/components/atoms/search-bar/search-bar.tsx
