+++ frontend/src/components/templates/homepage-header.tsx
@@ -39,6 +39,7 @@ export default function HomepageHeader() {
 
   // state of the scroll position and header height
   const [scrollPosition, setScrollPosition] = useState(0);
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const headerRef = useRef < any > (null);
   const [headerHeight, setHeaderHeight] = useState(20);
 
@@ -103,7 +104,7 @@ export default function HomepageHeader() {
     }
   }, [headerRef.current]);
 
-
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const onGitHubLoginSuccess = (response: any) => {
 
     setIsShowAuthModalOpen(false)
@@ -136,6 +137,7 @@ export default function HomepageHeader() {
 
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const onGitHubLoginFailure = (response: any) => {
     console.log('------the data from github-----failed-----', response);
   }
@@ -148,7 +150,7 @@ export default function HomepageHeader() {
         const provider = new WsProvider('wss://rpc.polkadot.io');
         const polkadotAPI = await ApiPromise.create({ provider });
         const address = accounts[0].address;
-        const data = await polkadotAPI.query.system.account(address);
+        await polkadotAPI.query.system.account(address);
       } catch (error) {
         console.error('Error connecting to wallet:', error);
       }
@@ -242,147 +244,9 @@ export default function HomepageHeader() {
               </div>
             </div>
           </div>
-          <div className="absolute right-0 top-0 z-[-1] opacity-30 lg:opacity-100">
-            <svg
-              width="450"
-              height="556"
-              viewBox="0 0 450 556"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <circle
-                cx="277"
-                cy="63"
-                r="225"
-                fill="url(#paint0_linear_25:217)"
-              />
-              <circle
-                cx="17.9997"
-                cy="182"
-                r="18"
-                fill="url(#paint1_radial_25:217)"
-              />
-              <circle
-                cx="76.9997"
-                cy="288"
-                r="34"
-                fill="url(#paint2_radial_25:217)"
-              />
-              <circle
-                cx="325.486"
-                cy="302.87"
-                r="180"
-                transform="rotate(-37.6852 325.486 302.87)"
-                fill="url(#paint3_linear_25:217)"
-              />
-              <circle
-                opacity="0.8"
-                cx="184.521"
-                cy="315.521"
-                r="132.862"
-                transform="rotate(114.874 184.521 315.521)"
-                stroke="url(#paint4_linear_25:217)"
-              />
-              <circle
-                opacity="0.8"
-                cx="356"
-                cy="290"
-                r="179.5"
-                transform="rotate(-30 356 290)"
-                stroke="url(#paint5_linear_25:217)"
-              />
-              <circle
-                opacity="0.8"
-                cx="191.659"
-                cy="302.659"
-                r="133.362"
-                transform="rotate(133.319 191.659 302.659)"
-                fill="url(#paint6_linear_25:217)"
-              />
-              <defs>
-                <linearGradient
-                  id="paint0_linear_25:217"
-                  x1="-54.5003"
-                  y1="-178"
-                  x2="222"
-                  y2="288"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="#4A6CF7" />
-                  <stop offset="1" stopColor="#4A6CF7" stopOpacity="0" />
-                </linearGradient>
-                <radialGradient
-                  id="paint1_radial_25:217"
-                  cx="0"
-                  cy="0"
-                  r="1"
-                  gradientUnits="userSpaceOnUse"
-                  gradientTransform="translate(17.9997 182) rotate(90) scale(18)"
-                >
-                  <stop offset="0.145833" stopColor="#4A6CF7" stopOpacity="0" />
-                  <stop offset="1" stopColor="#4A6CF7" stopOpacity="0.08" />
-                </radialGradient>
-                <radialGradient
-                  id="paint2_radial_25:217"
-                  cx="0"
-                  cy="0"
-                  r="1"
-                  gradientUnits="userSpaceOnUse"
-                  gradientTransform="translate(76.9997 288) rotate(90) scale(34)"
-                >
-                  <stop offset="0.145833" stopColor="#4A6CF7" stopOpacity="0" />
-                  <stop offset="1" stopColor="#4A6CF7" stopOpacity="0.08" />
-                </radialGradient>
-                <linearGradient
-                  id="paint3_linear_25:217"
-                  x1="226.775"
-                  y1="-66.1548"
-                  x2="292.157"
-                  y2="351.421"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="#4A6CF7" />
-                  <stop offset="1" stopColor="#4A6CF7" stopOpacity="0" />
-                </linearGradient>
-                <linearGradient
-                  id="paint4_linear_25:217"
-                  x1="184.521"
-                  y1="182.159"
-                  x2="184.521"
-                  y2="448.882"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="#4A6CF7" />
-                  <stop offset="1" stopColor="white" stopOpacity="0" />
-                </linearGradient>
-                <linearGradient
-                  id="paint5_linear_25:217"
-                  x1="356"
-                  y1="110"
-                  x2="356"
-                  y2="470"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="#4A6CF7" />
-                  <stop offset="1" stopColor="white" stopOpacity="0" />
-                </linearGradient>
-                <linearGradient
-                  id="paint6_linear_25:217"
-                  x1="118.524"
-                  y1="29.2497"
-                  x2="166.965"
-                  y2="338.63"
-                  gradientUnits="userSpaceOnUse"
-                >
-                  <stop stopColor="#4A6CF7" />
-                  <stop offset="1" stopColor="#4A6CF7" stopOpacity="0" />
-                </linearGradient>
-              </defs>
-            </svg>
-          </div>
           <div className="absolute bottom-0 left-0 z-[-1] opacity-30 lg:opacity-100">
             <svg
-              width="364"
+              width="full"
               height="201"
               viewBox="0 0 364 201"
               fill="none"
@@ -573,7 +437,7 @@ export default function HomepageHeader() {
                                   }}
                                 >
                                   {chain.iconUrl && (
-                                    <img
+                                    <Image
                                       alt={chain.name ?? 'Chain icon'}
                                       src={chain.iconUrl}
                                       style={{ width: 12, height: 12 }}
